feat: Pockets as universal canvas with Mission Control + Deep Work types
Pockets replace the command center as the primary dashboard surface.
Everything is a Pocket — Mission Control, Deep Work, Sprint Tracker,
Research Hub, Content Pipeline are all pocket types.

Changes:
- Add /pockets route (replaces /command-center, redirect kept)
- Add Mission Control pocket (agent roster, task kanban, activity feed)
- Add Deep Work pocket (project kanban, standup, deliverables)
- Register both in kit catalog as built-in pocket types
- Update WorkspaceTabs nav: "Pockets" tab enabled, links to /pockets
- Rename "Kit Store" to "Pocket Store" in UI
