+++ src/pocketpaw/kits/builtins/mission_control.yaml
@@ -1,0 +1,60 @@
+# Mission Control Pocket — multi-agent orchestration dashboard.
+# Created: 2026-03-21
+meta:
+  name: "Mission Control"
+  author: "PocketPaw"
+  version: "1.0.0"
+  description: "Monitor agents, delegate tasks, and track multi-agent workflows"
+  category: general
+  icon: radio-tower
+  tags: [agents, orchestration, mission-control, multi-agent]
+  built_in: true
+
+layout:
+  columns: 2
+  sections:
+    - title: "Fleet Status"
+      span: full
+      panels:
+        - id: metrics
+          type: metrics-row
+          items:
+            - { label: "Active Agents", source: "api:stats", field: "agents.active", format: "number" }
+            - { label: "Tasks In Progress", source: "api:stats", field: "tasks.by_status.in_progress", format: "number" }
+            - { label: "Completed Today", source: "api:stats", field: "tasks.by_status.done", format: "number" }
+            - { label: "Blocked", source: "api:stats", field: "tasks.by_status.blocked", format: "number" }
+
+    - title: "Agent Roster"
+      span: left
+      panels:
+        - id: agents
+          type: table
+          source: "api:agents"
+          columns:
+            - { key: "name", label: "Agent" }
+            - { key: "role", label: "Role" }
+            - { key: "status", label: "Status" }
+            - { key: "current_task_id", label: "Current Task" }
+
+    - title: "Task Board"
+      span: right
+      panels:
+        - id: kanban
+          type: kanban
+          columns:
+            - { key: "assigned", label: "Assigned", color: "gray" }
+            - { key: "in_progress", label: "Working", color: "blue" }
+            - { key: "review", label: "Review", color: "orange" }
+            - { key: "done", label: "Done", color: "green" }
+          source: "api:tasks"
+          card_fields: [title, priority, assignee_ids]
+
+    - title: "Activity Feed"
+      span: full
+      panels:
+        - id: activity
+          type: feed
+          source: "api:activities"
+          max_items: 50
+
+workflows: {}
